Add files via upload
The Titanic Prediction project is a classic machine learning task that involves predicting whether a passenger on the Titanic survived or not based on various features. This project is commonly used as a beginner-friendly introduction to data analysis, feature engineering, and machine learning classification algorithms.

ools Used:

    Jupyter Notebook: An interactive computing environment for data analysis and visualization.
    Python Libraries: Pandas, NumPy, Matplotlib, Seaborn for data manipulation, analysis, and visualization.
    Scikit-learn: A machine learning library for building and evaluating predictive models.

Learning Objectives:

    Gain hands-on experience in data analysis and preprocessing.
    Understand the workflow of building a predictive model.
    Learn how to use machine learning algorithms for classification tasks.
    Improve skills in data visualization and interpretation.

By completing the Titanic Prediction project in Jupyter Notebook, participants will develop a foundational understanding of data science concepts and be better equipped to tackle more complex machine learning projects in the future.